refactor: separate commands into modules (#1254)
## Proposed Changes

<!-- Describe the changes proposed in this pull request. -->
<!-- Please provide links to any issue(s) which are expected to be
resolved. -->
This makes no (intentional) behavioral changes. The changes are:
- Separate the commands that were previously in
`commands/environment.rs` into separate modules.
- Renames `EnvironmentError2` -> `EnvironmentError`
- Renames the test files from `environment-*.bats` to `*.bats` since
they are no longer under the `environment` module anymore (and we
already had a mix of `environment-*` and `*` files.

3 integration tests were slightly changed in that their regular
expressions _only_ allowed versions with three components e.g. `2.12.1`.
Some package versions only contain two components e.g. `2.12`. This
change makes the regular expression support either case.

## Release Notes

<!-- Describe any user facing changes. Use "N/A" if not applicable. -->
N/A

<!-- Many thanks! -->
