+++ cli/tests/activate.bats
