+++ cli/tests/containerize.bats
