+++ cli/tests/delete.bats
