+++ cli/tests/edit.bats
