+++ cli/tests/init.bats
