+++ cli/tests/install.bats
