+++ cli/tests/pull.bats
