+++ cli/tests/search.bats
