+++ cli/tests/show.bats
